requirements.txt
mcp 
mcp-ui-server
fastmcp
pandas
requests
python-dotenv
